Move Individual data population to startupViewModel
