+++ src/main/res/layout/about_activity.xml
@@ -90,21 +90,6 @@
             tools:layout_conversion_absoluteY="205dp"/>
 
         <Button
-            android:id="@+id/restTestButton2"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:text="Test Individual REST Call"
-            android:onClick="@{() -> viewModel.testIndividualServiceCall()}"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/createDatabaseButton"
-            tools:layout_conversion_absoluteHeight="48dp"
-            tools:layout_conversion_absoluteWidth="131dp"
-            tools:layout_conversion_absoluteX="126dp"
-            tools:layout_conversion_absoluteY="205dp"/>
-
-        <Button
             android:id="@+id/textTableChangeButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
